refactor(config): add config to core folder
