--- src/config/docs/index.ts
@@ -1,13 +1,0 @@
-import type { INestApplication } from '@nestjs/common';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-
-export const setupSwagger = (app: INestApplication) => {
-	const config = new DocumentBuilder()
-		.setTitle('API Projeto de Extensão ')
-		.setDescription('API do meu Projeto de extensão')
-		.setVersion('1.0.0')
-		.build();
-
-	const document = () => SwaggerModule.createDocument(app, config);
-	SwaggerModule.setup('api', app, document);
-};
